refactor: move previous/next buttons into SelectionView
